+++ src/commands/delete.ts
@@ -1,8 +1,9 @@
 import * as vscode from 'vscode';
 import * as yaml from '../yaml';
 import * as kube from '../kube';
+import { ObjectNode, TreeDataProvider } from '../TreeDataProvider';
 
-export async function deleteObjectFromActiveEditor() {
+export async function deleteObjectFromActiveEditor(treeDataProvider: TreeDataProvider) {
 	let editor = vscode.window.activeTextEditor;
 	if (!editor) {
 		return;
@@ -14,4 +15,9 @@ export async function deleteObjectFromActiveEditor() {
 
 	await kube.api.delete(kube.api.getObjectUri(obj));
   await vscode.commands.executeCommand('workbench.action.closeActiveEditor');
+
+	if (typeof obj === 'object' && obj.metadata) {
+		let parent = new ObjectNode(obj).getParent();
+		treeDataProvider.invalidate(parent);
+	}
 }